+++ pkg/thumbnail/imgsource/webdav.go
