+++ pkg/thumbnail/storage/filesystem.go
