+++ pkg/thumbnail/storage/inmemory.go
