+++ pkg/thumbnail/storage/storage.go
@@ -1,12 +1,12 @@
 package storage
 
-import "github.com/owncloud/ocis-thumbnails/pkg/thumbnails/resolutions"
+import "github.com/owncloud/ocis-thumbnails/pkg/thumbnail/resolution"
 
 // Request combines different attributes needed for storage operations.
 type Request struct {
 	ETag       string
 	Types      []string
-	Resolution resolutions.Resolution
+	Resolution resolution.Resolution
 }
 
 // Storage defines the interface for a thumbnail store.
